successfully modularized the entire app
