+++ src/javascripts/helpers/data/bears.js
@@ -1,0 +1,21 @@
+const bears = [
+  {
+    name: 'Flash',
+    imageUrl: 'https://images.pexels.com/photos/35435/pexels-photo.jpg?cs=srgb&dl=pexels-rasmus-svinding-35435.jpg&fm=jpg',
+  },
+  {
+    name: 'Mary',
+    imageUrl: 'https://images.pexels.com/photos/34700/bear-animals-zoo-captivity.jpg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
+  },
+  {
+    name: 'Bill',
+    imageUrl: 'https://i.pinimg.com/236x/90/d5/c2/90d5c2c2fdfbd3eab4c4e4a2190d0a88.jpg',
+  },
+  {
+    name: 'Louise',
+    imageUrl: 'https://www.publicdomainpictures.net/pictures/230000/velka/brown-bear-1504708649cqE.jpg',
+  },
+
+];
+
+export default bears;
