+++ src/javascripts/helpers/domEvents.js
@@ -1,0 +1,7 @@
+import getFormInfo from '../components/getFormInfo';
+
+const domEvents = () => {
+  document.querySelector('form').addEventListener('submit', getFormInfo);
+};
+
+export default domEvents;
